src/database-api/requirement.txt
fastapi
uvicorn
mysql-connector-python
statsmodels
pandas
scikit-learn